GCNexperiments/requirements.txt
pytorch
pytorch_geometric
numpy
wandb
seaborn
scikit-learn